ci_api/requirements.txt
alembic==1.8.1
asyncpg==0.27.0
fastapi==0.85.1
fastapi-mail==1.2.0
pyjwt==2.6.0
loguru==0.6.0
passlib==1.7.4
pydantic[email]
pytest-asyncio==0.20.2
python-dotenv==0.21.0
python-multipart==0.0.5
sqladmin==0.7.0
sqlmodel==0.0.8
uvicorn==0.19.0
